Close audit findings: security + CVE + infra + deal-detection + refactor
Closes ~130 of the 134 findings from docs/AUDIT_REPORT.md. Three
high-risk refactors (snapshot_farvater split, bot mypy cleanup, full
shared/infra adoption everywhere) are deferred to follow-up PRs with
specs in docs/superpowers/specs/.

=== P0 SECURITY (audit #1-9 critical/high) ===
* HMAC compare_digest in /alerts webhook, fail-closed in prod
* assert_prod_secrets requires ALERTMANAGER_WEBHOOK_SECRET
* alertmanager port closed in prod overlay (ports !reset [])
* Per-IP rate-limit keying via CF-Connecting-IP -> XFF -> client.host
* CVE bumps: aiogram 3.25 (drops 19 aiohttp CVEs), starlette 0.49.3,
  python-multipart 0.0.29, urllib3 2.7, curl-cffi 0.15, fastapi 0.132
* R2 backup encryption via age; ssh-keyscan TOFU removed (known_hosts
  now required); pg_dump switched to -Fc for parallel restore
* Compose guardrails: json-file logging max-size, mem_limit per service,
  Prometheus retention 30d/8GB, Grafana admin password required (no
  admin:admin fallback)
* Bot pyproject.toml + multi-stage Dockerfile + non-root USER +
  HEALTHCHECK; bot now tracked by Dependabot pip ecosystem

=== DEAL DETECTION (main product goal) ===
* Cold-start tagged 'peer_anomaly' instead of 'percentile' so it can
  be differentiated downstream
* Channel broadcast (post_deals) excludes peer_anomaly and orders by
  discount_pct DESC instead of detected_at DESC
* notify_subscribers requires 25% floor for peer_anomaly alerts; picks
  the deepest discount per filter per tick
* API /api/deals: default sort=discount, expose detection_method
* Bot templates show savings in UAH, strike-through baseline,
  per-method 'why this is cheap' explanation
* /best command for top-N deepest discounts

=== INFRA / OBSERVABILITY ===
* node + postgres + redis exporters with 4 host-level alerts
  (DiskFull80, OOMRiskHigh, PostgresDeadTuplesHigh, RedisDown)
* Certbot deploy hook + certbot.timer enabled in cloud-init
* unattended-upgrades for OS security patches
* Loki + Vector log aggregation under 'observability' compose profile
* Terraform R2 backend (operator runs terraform init -backend-config)
* All 12 GitHub Actions pinned to commit SHAs (audit Quarter #20)
* Pre-commit config (ruff, prettier, shellcheck, hadolint)
* SBOM via anchore/sbom-action in deploy-api workflow
* Deploy waits for CI via workflow_run trigger (no more red-CI/green-deploy)
* :latest tags dropped from image pushes (sha-XYZ only for rollback safety)

=== REFACTORS ===
* apps/shared/infra/ — BaseAppSettings, configure_logging,
  configure_sentry, get_redis_factory (audit Sprint #7); services
  inherit instead of copy-pasting
* routers/hotels.py split: refresh queue moved to services/refresh_queue.py
* detect_deals uses strategy table instead of 5-deep budget arithmetic
* /api/hotels/{id}/calendar no longer mutates Redis on GET (replaced
  with explicit POST /hot-ping, rate-limited 60/min/IP)
* AmpQueryParamMiddleware rewrites ?amp;param= globally; search router
  no longer has 12 alias Query() declarations

=== POSTGRES ===
* pg_stat_statements loaded, max_connections=200
* WAL archiving config + pg_wal_archive.sh script (off by default;
  operator enables after wiring R2)
* SSL config templated (commented; operator uncomments for prod)
* PostGIS migration 018 adds hotels.coords_geo + GIST index, backfills
  from legacy text column

=== FRONTEND ===
* React 19 RC -> 19.2 stable; @types/react 18 -> 19
* CSP, HSTS, Permissions-Policy headers in next.config.mjs
* Vitest config now matches .tsx; jsdom + React Testing Library
* DealCard.test.tsx covers headline contract (5 cases)
* Deal type carries detection_method; api-client accepts sort param

=== QUALITY GATES ===
* pytest-cov in dev-deps for api/bot/scheduler/ingest; CI emits coverage
* mypy ratchet (infra/scripts/mypy-ratchet.sh) with baseline at 0/0/0/70
  — strict=true now meaningfully enforced; new code can't regress typing
* scheduler CI gains postgres+redis services for upcoming integration tests
* Hotel.review_score typed Decimal (was incorrectly float)
* Bandit MD5 noise silenced via usedforsecurity=False
* hadolint DL3008 ignores documented per Dockerfile
* secrets-bootstrap.sh uses openssl rand -hex (was base64+tr alphabet shrink)

=== DOCS ===
* docs/AUDIT_REPORT.md: the source of all this work
* docs/superpowers/specs/2026-05-27-snapshot-farvater-split.md:
  deferred plan for the 1284-LOC monolith (concrete 6-step migration)
* Stale docs (ARCHITECTURE, DECISIONS, INFRASTRUCTURE, TESTING, etc)
  removed; AUDIT_REPORT is the new canonical reference
* .secrets/README.md (gitignored) documents the age keypair operator
  flow and lists the public key to put in GitHub secrets

=== VERIFICATION (this commit) ===
* ruff: clean across api/scheduler/ingest/bot (4/4 services)
* mypy ratchet: api=0, scheduler=0, ingest=0, bot=70 (locked baseline)
* tests: bot 43/43, scheduler 139/139, api unit 7/7, web vitest 9/9
* promtool: 9/9 alert rules valid
* docker compose: dev + prod overlay + observability profile all validate

=== OPERATOR TODO ===
1. Set GH secret BACKUP_AGE_RECIPIENT to the age public key in
   .secrets/README.md (then move private key out of .secrets/)
2. Capture ssh-keyscan -H <vps> into DEPLOY_SSH_KNOWN_HOSTS +
   BACKUP_SSH_KNOWN_HOSTS secrets
3. Rotate TELEGRAM_BOT_TOKEN + ALERTMANAGER_WEBHOOK_SECRET (audit #1)
4. Create R2 bucket fasttravel-tfstate + token, then
   `terraform init -migrate-state -backend-config=...`

+++ apps/api/src/models/hotel.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING, Any
 
 from sqlalchemy import (
@@ -53,7 +54,13 @@ class Hotel(Base):
     description_uk: Mapped[str | None] = mapped_column(Text, nullable=True)
     photos_jsonb: Mapped[list[dict[str, Any]] | None] = mapped_column(JSONB, nullable=True)
     amenities: Mapped[list[str] | None] = mapped_column(ARRAY(Text), nullable=True)
-    review_score: Mapped[float | None] = mapped_column(Numeric(3, 1), nullable=True)
+    # Audit 1.3 Low — SQLAlchemy returns Numeric columns as Decimal, not
+    # float. Annotation was a quiet lie; consumers doing
+    # `float(hotel.review_score)` worked but consumers doing arithmetic
+    # without coercion silently mixed Decimal + float and lost
+    # precision. Now the type matches the runtime; convert at the
+    # serialization boundary (`float(...)` in templates / pydantic).
+    review_score: Mapped[Decimal | None] = mapped_column(Numeric(3, 1), nullable=True)
     review_count: Mapped[int] = mapped_column(Integer, nullable=False, server_default="0")
     last_updated: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default="true")
